Allow the lists to include an asterisk to be able to distinguish between
1998-09-28  Miguel de Icaza  <miguel@nuclecu.unam.mx>

	* src/sheet-autofill.c (matches_list): Allow the lists to include
	an asterisk to be able to distinguish between abbreviations and
	full descriptions.

	* src/clipboard.c (x_selection_to_cell_region): New routine.
	Creates a CellRegion of type text based on the selection provided
	by X.
	(paste_cell_flags): New function, decoupled from
	do_clipboard_paste_cell_region.   Handles new type in CellCopy.

	* src/sheet.c (sheet_selection_paste): Simplified as most of the
	work is now

	* src/clipboard.c (x_selection_received): Perform the paste
	operation, moved most of the logic from sheet.c.  This does both
	internal paste and paste from the X selection (and everything
	works the way the user expects it to behave).

+++ src/cell.c
@@ -84,7 +84,7 @@ cell_set_halign (Cell *cell, StyleHAlignFlags halign)
 {
 	g_return_if_fail (cell != NULL);
 
-	if (cell->style->halign == (unsigned int) halign)
+	if (((unsigned int)cell->style->halign) == ((unsigned int) halign))
 		return;
 
 	cell_queue_redraw (cell);
